Change Home page to Explore page Added login logic to the router Added vercel.json config Small general fixes
