+++ frontend/src/components/explore/ui/carousel/ExploreCarousel.vue
